requirements.txt
# Crew Fatigue Intelligence System — Backend Requirements
# Core
numpy>=1.26.0
pandas>=2.1.0
scikit-learn>=1.4.0
xgboost>=2.0.0
shap>=0.44.0
matplotlib>=3.8.0
scipy>=1.11.0

# API
fastapi>=0.109.0
uvicorn[standard]>=0.27.0
pydantic>=2.0.0
python-dotenv>=1.0.0

# MongoDB
motor>=3.3.0
pymongo>=4.6.0